+++ src/components/Grid/style.scss
@@ -15,10 +15,12 @@
 }
 
 .mapCoordsGrid {
-  z-index: -100;
+  z-index: -1;
   position: absolute;
   top: 0;
   left: 0;
+  right: 0;
+  bottom: 0;
   font-size: small;
   color: #aaa;
   text-align: left;
