+++ src/components/SliderDrawer/index.js
@@ -39,6 +39,7 @@ const useStyles = makeStyles((theme) => ({
     marginLeft: 25,
     marginTop:25,
     paddingTop: 25,
+    paddingBottom: 100,
     display: 'flex',
     flexWrap: 'wrap',
     width: '100%',
@@ -77,7 +78,7 @@ export default function SliderDrawer({ handleDraggableItem }) {
   const [state, setState] = React.useState({
     isDrawerOpened: false
   })
-  const [tileList, setTileList] = React.useState([]);
+  const [tileSet, setTileSet] = React.useState([]);
 
   const handleDrawerOpen = () => {
     setState({
@@ -93,11 +94,29 @@ export default function SliderDrawer({ handleDraggableItem }) {
 
   const { isDrawerOpened } = state;
 
+  /**
+   * <DraggableTile key="0" tileId="0" environment="swamp" imageURL="https://picsum.photos/seed/crocodile/100" handleOnDragStart={handleDraggableItem} />
+   * key, tileId, environment (name as string), imgURL (for bg)
+   */
+
   React.useEffect(() => {
     API.getTiles()
     .then(tiles => {
-      console.log(tiles);
+      const list = tiles.data;
+      let tileList = [];
+      for(let i = 0; i < list.length; i++) {
+        const tile = {
+          key: i,
+          tileId: list[i].id,
+          environment: list[i].Environment.name,
+          imageURL:  list[i].image_url
+        }
+        tileList.push(tile);
+      }
+      // console.log(tileList)
+      setTileSet(tileList);
     })
+    .catch(err => console.error(err))
   },[]);
   
   return (
@@ -126,10 +145,7 @@ export default function SliderDrawer({ handleDraggableItem }) {
         {/* Render the top 18 until scroll down, then render more, etc */}
         <div className={classes.tileGrid}>
           {/* Set this as {children} to handle whether its nav or tiles */}
-          <DraggableTile key="0" tileId="0" environment="swamp" imageURL="https://picsum.photos/seed/crocodile/100" handleOnDragStart={handleDraggableItem} />
-          <DraggableTile key="1" tileId="1" environment="swamp" imageURL="https://picsum.photos/seed/alligator/100" handleOnDragStart={handleDraggableItem} />
-          <DraggableTile key="2" tileId="2" environment="swamp" imageURL="https://picsum.photos/seed/gharial/100" handleOnDragStart={handleDraggableItem} />
-          <DraggableTile key="3" tileId="3" environment="swamp" imageURL="https://picsum.photos/seed/flamingo/100" handleOnDragStart={handleDraggableItem} />
+          {tileSet.map(tile => <DraggableTile key={tile.key} tileId={tile.tileId} environment={tile.environment} imageURL={tile.imageURL} handleOnDragStart={handleDraggableItem} />)}
         </div>
       </TileDrawer>
     </div>
